+++ src/fidesops/ops/email_templates/get_email_template.py
@@ -6,6 +6,8 @@ from jinja2 import Environment, FileSystemLoader, Template, select_autoescape
 from fidesops.ops.common_exceptions import EmailTemplateUnhandledActionType
 from fidesops.ops.email_templates.template_names import (
     EMAIL_ERASURE_REQUEST_FULFILLMENT,
+    PRIVACY_REQUEST_COMPLETE_ACCESS_TEMPLATE,
+    PRIVACY_REQUEST_COMPLETE_DELETION_TEMPLATE,
     SUBJECT_IDENTITY_VERIFICATION_TEMPLATE,
 )
 from fidesops.ops.schemas.email.email import EmailActionType
@@ -25,6 +27,10 @@ def get_email_template(action_type: EmailActionType) -> Template:
         return template_env.get_template(SUBJECT_IDENTITY_VERIFICATION_TEMPLATE)
     if action_type == EmailActionType.EMAIL_ERASURE_REQUEST_FULFILLMENT:
         return template_env.get_template(EMAIL_ERASURE_REQUEST_FULFILLMENT)
+    if action_type == EmailActionType.PRIVACY_REQUEST_COMPLETE_ACCESS:
+        return template_env.get_template(PRIVACY_REQUEST_COMPLETE_ACCESS_TEMPLATE)
+    if action_type == EmailActionType.PRIVACY_REQUEST_COMPLETE_DELETION:
+        return template_env.get_template(PRIVACY_REQUEST_COMPLETE_DELETION_TEMPLATE)
 
     logger.error("No corresponding template linked to the %s", action_type)
     raise EmailTemplateUnhandledActionType(