+++ src/fidesops/ops/service/email/email_dispatch_service.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Any, Optional, Union
+from typing import Any, Optional
 
 import requests
 from requests import Response
@@ -8,14 +8,12 @@ from sqlalchemy.orm import Session
 from fidesops.ops.common_exceptions import EmailDispatchException
 from fidesops.ops.email_templates import get_email_template
 from fidesops.ops.models.email import EmailConfig
-from fidesops.ops.models.privacy_request import EmailRequestFulfillmentBodyParams
 from fidesops.ops.schemas.email.email import (
     EmailActionType,
     EmailForActionType,
     EmailServiceDetails,
     EmailServiceSecrets,
     EmailServiceType,
-    SubjectIdentityVerificationBodyParams,
 )
 from fidesops.ops.util.logger import Pii
 
@@ -26,10 +24,7 @@ def dispatch_email(
     db: Session,
     action_type: EmailActionType,
     to_email: Optional[str],
-    email_body_params: Union[
-        SubjectIdentityVerificationBodyParams,
-        EmailRequestFulfillmentBodyParams,
-    ],
+    email_body_params: Any,
 ) -> None:
     if not to_email:
         raise EmailDispatchException("No email supplied.")
@@ -81,6 +76,22 @@ def _build_email(
             body=base_template.render(
                 {"dataset_collection_action_required": body_params}
             ),
+        )
+    if action_type == EmailActionType.PRIVACY_REQUEST_COMPLETE_ACCESS:
+        base_template = get_email_template(action_type)
+        return EmailForActionType(
+            subject="Your data is ready to be downloaded",
+            body=base_template.render(
+                {
+                    "download_links": body_params.download_links,
+                }
+            ),
+        )
+    if action_type == EmailActionType.PRIVACY_REQUEST_COMPLETE_DELETION:
+        base_template = get_email_template(action_type)
+        return EmailForActionType(
+            subject="Your data has been deleted",
+            body=base_template.render(),
         )
     logger.error("Email action type %s is not implemented", action_type)
     raise EmailDispatchException(f"Email action type {action_type} is not implemented")