+++ src/fidesops/ops/service/privacy_request/request_runner_service.py
@@ -1,7 +1,7 @@
 import logging
 import random
 from datetime import datetime, timedelta
-from typing import ContextManager, Dict, List, Optional, Set
+from typing import Any, ContextManager, Dict, List, Optional, Set
 
 from celery import Task
 from celery.utils.log import get_task_logger
@@ -15,6 +15,7 @@ from fidesops.ops import common_exceptions
 from fidesops.ops.common_exceptions import (
     ClientUnsuccessfulException,
     EmailDispatchException,
+    IdentityNotFoundException,
     PrivacyRequestPaused,
 )
 from fidesops.ops.core.config import config
@@ -36,9 +37,15 @@ from fidesops.ops.models.policy import (
 from fidesops.ops.models.privacy_request import (
     PrivacyRequest,
     PrivacyRequestStatus,
+    ProvidedIdentityType,
     can_run_checkpoint,
 )
+from fidesops.ops.schemas.email.email import (
+    AccessRequestCompleteBodyParams,
+    EmailActionType,
+)
 from fidesops.ops.service.connectors.email_connector import email_connector_erasure_send
+from fidesops.ops.service.email.email_dispatch_service import dispatch_email
 from fidesops.ops.service.storage.storage_uploader_service import upload
 from fidesops.ops.task.filter_results import filter_data_categories
 from fidesops.ops.task.graph_task import (
@@ -123,11 +130,11 @@ def upload_access_results(
     access_result: Dict[str, List[Row]],
     dataset_graph: DatasetGraph,
     privacy_request: PrivacyRequest,
-) -> None:
+) -> List[str]:
     """Process the data uploads after the access portion of the privacy request has completed"""
+    download_urls: List[str] = []
     if not access_result:
         logging.info("No results returned for access request %s", privacy_request.id)
-
     for rule in policy.get_rules_for_action(action_type=ActionType.access):
         if not rule.storage_destination:
             raise common_exceptions.RuleValidationError(
@@ -145,12 +152,14 @@ def upload_access_results(
             privacy_request.id,
         )
         try:
-            upload(
+            download_url: Optional[str] = upload(
                 db=session,
                 request_id=privacy_request.id,
                 data=filtered_results,
                 storage_key=rule.storage_destination.key,  # type: ignore
             )
+            if download_url:
+                download_urls.append(download_url)
         except common_exceptions.StorageUploadError as exc:
             logging.error(
                 "Error uploading subject access data for rule %s on policy %s and privacy request %s : %s",
@@ -160,6 +169,7 @@ def upload_access_results(
                 Pii(str(exc)),
             )
             privacy_request.status = PrivacyRequestStatus.error
+    return download_urls
 
 
 def queue_privacy_request(
@@ -168,6 +178,7 @@ def queue_privacy_request(
     from_step: Optional[str] = None,
 ) -> str:
     cache: FidesopsRedis = get_cache()
+    logger.info("queueing privacy request")
     task = run_privacy_request.delay(
         privacy_request_id=privacy_request_id,
         from_webhook_id=from_webhook_id,
@@ -207,7 +218,7 @@ async def run_privacy_request(
     from_webhook_id: Optional[str] = None,
     from_step: Optional[str] = None,
 ) -> None:
-    # pylint: disable=too-many-locals, too-many-statements
+    # pylint: disable=too-many-locals, too-many-statements, too-many-return-statements, too-many-branches
     """
     Dispatch a privacy_request into the execution layer by:
         1. Generate a graph from all the currently configured datasets
@@ -260,6 +271,7 @@ async def run_privacy_request(
             dataset_graph = DatasetGraph(*dataset_graphs)
             identity_data = privacy_request.get_cached_identity_data()
             connection_configs = ConnectionConfig.all(db=session)
+            access_result_urls: List[str] = []
 
             if can_run_checkpoint(
                 request_checkpoint=CurrentStep.access, from_checkpoint=resume_step
@@ -272,8 +284,7 @@ async def run_privacy_request(
                     identity=identity_data,
                     session=session,
                 )
-
-                upload_access_results(
+                access_result_urls = upload_access_results(
                     session,
                     policy,
                     access_result,
@@ -343,7 +354,19 @@ async def run_privacy_request(
         )
         if not proceed:
             return
-
+        if config.notifications.send_request_completion_notification:
+            try:
+                initiate_privacy_request_completion_email(
+                    session, policy, access_result_urls, identity_data
+                )
+            except (IdentityNotFoundException, EmailDispatchException) as e:
+                privacy_request.error_processing(db=session)
+                # If dev mode, log traceback
+                await fideslog_graph_failure(
+                    failed_graph_analytics_event(privacy_request, e)
+                )
+                _log_exception(e, config.dev_mode)
+                return
         privacy_request.finished_processing_at = datetime.utcnow()
         AuditLog.create(
             db=session,
@@ -355,8 +378,42 @@ async def run_privacy_request(
             },
         )
         privacy_request.status = PrivacyRequestStatus.complete
-        privacy_request.save(db=session)
         logging.info("Privacy request %s run completed.", privacy_request.id)
+        privacy_request.save(db=session)
+
+
+def initiate_privacy_request_completion_email(
+    session: Session,
+    policy: Policy,
+    access_result_urls: List[str],
+    identity_data: Dict[str, Any],
+) -> None:
+    """
+    :param session: SQLAlchemy Session
+    :param policy: Policy
+    :param access_result_urls: list of urls generated by access request upload
+    :param identity_data: Dict of identity data
+    """
+    if not identity_data.get(ProvidedIdentityType.email.value):
+        raise IdentityNotFoundException(
+            "Identity email was not found, so request completion email could not be sent."
+        )
+    if policy.get_rules_for_action(action_type=ActionType.access):
+        dispatch_email(
+            db=session,
+            action_type=EmailActionType.PRIVACY_REQUEST_COMPLETE_ACCESS,
+            to_email=identity_data.get(ProvidedIdentityType.email.value),
+            email_body_params=AccessRequestCompleteBodyParams(
+                download_links=access_result_urls
+            ),
+        )
+    if policy.get_rules_for_action(action_type=ActionType.erasure):
+        dispatch_email(
+            db=session,
+            action_type=EmailActionType.PRIVACY_REQUEST_COMPLETE_DELETION,
+            to_email=identity_data.get(ProvidedIdentityType.email.value),
+            email_body_params=None,
+        )
 
 
 def initiate_paused_privacy_request_followup(privacy_request: PrivacyRequest) -> None: