+++ src/fidesops/ops/service/storage/storage_authenticator_service.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Any, Dict
+from typing import Any
 
 from botocore.exceptions import ClientError
 from requests import RequestException
@@ -7,8 +7,8 @@ from requests import RequestException
 from fidesops.ops.schemas.storage.storage import (
     SUPPORTED_STORAGE_SECRETS,
     S3AuthMethod,
-    StorageSecrets,
     StorageSecretsOnetrust,
+    StorageSecretsS3,
     StorageType,
 )
 from fidesops.ops.util.storage_authenticator import (
@@ -28,10 +28,10 @@ def secrets_are_valid(
     return uploader(secrets)
 
 
-def _s3_authenticator(secrets: Dict[StorageSecrets, Any]) -> bool:
+def _s3_authenticator(secrets: StorageSecretsS3) -> bool:
     """Authenticates secrets for s3, returns true if secrets are valid"""
     try:
-        get_s3_session(S3AuthMethod.SECRET_KEYS, secrets)
+        get_s3_session(S3AuthMethod.SECRET_KEYS.value, secrets.dict())  # type: ignore
         return True
     except ClientError:
         return False