+++ src/fidesops/ops/tasks/storage.py
@@ -11,6 +11,7 @@ from typing import Any, Dict, Union
 
 import pandas as pd
 import requests
+from boto3 import Session
 from botocore.exceptions import ClientError, ParamValidationError
 from fideslib.cryptography.cryptographic_util import bytes_to_b64_str
 
@@ -27,7 +28,6 @@ from fidesops.ops.util.storage_authenticator import (
 )
 
 logger = logging.getLogger(__name__)
-
 
 LOCAL_FIDES_UPLOAD_DIRECTORY = "fides_uploads"
 
@@ -96,6 +96,27 @@ def write_to_in_memory_buffer(
     raise NotImplementedError(f"No handling for response format {resp_format}.")
 
 
+def create_presigned_url_for_s3(
+    s3_client: Session, bucket_name: str, object_name: str
+) -> str:
+    """ "Generate a presigned URL to share an S3 object
+
+    :param s3_client: s3 base client
+    :param bucket_name: string
+    :param object_name: string
+    :return: Presigned URL as string.
+    """
+
+    response = s3_client.generate_presigned_url(
+        "get_object",
+        Params={"Bucket": bucket_name, "Key": object_name},
+        ExpiresIn=config.security.subject_request_download_link_ttl_seconds,
+    )
+
+    # The response contains the presigned URL
+    return response
+
+
 def upload_to_s3(  # pylint: disable=R0913
     storage_secrets: Dict[StorageSecrets, Any],
     data: Dict,
@@ -109,18 +130,28 @@ def upload_to_s3(  # pylint: disable=R0913
     logger.info("Starting S3 Upload of %s", file_key)
     try:
         my_session = get_s3_session(auth_method, storage_secrets)
-
-        s3 = my_session.client("s3")
+        s3_client = my_session.client("s3")
 
         # handles file chunking
-        s3.upload_fileobj(
-            Fileobj=write_to_in_memory_buffer(resp_format, data, request_id),
-            Bucket=bucket_name,
-            Key=file_key,
+        try:
+            s3_client.upload_fileobj(
+                Fileobj=write_to_in_memory_buffer(resp_format, data, request_id),
+                Bucket=bucket_name,
+                Key=file_key,
+            )
+        except Exception as e:
+            logger.error("Encountered error while uploading s3 object: %s", e)
+            raise e
+
+        presigned_url: str = create_presigned_url_for_s3(
+            s3_client, bucket_name, file_key
         )
-        # todo- move to outbound_urn_registry
-        return f"https://{bucket_name}.s3.amazonaws.com/{file_key}"
+
+        return presigned_url
     except ClientError as e:
+        logger.error(
+            "Encountered error while uploading and generating link for s3 object: %s", e
+        )
         raise e
     except ParamValidationError as e:
         raise ValueError(f"The parameters you provided are incorrect: {e}")
@@ -147,7 +178,7 @@ def upload_to_onetrust(
         data=payload,
         headers=headers,
     )
-    return "success"
+    return "onetrust"
 
 
 def _handle_json_encoding(field: Any) -> str:
@@ -169,4 +200,4 @@ def upload_to_local(payload: Dict, file_key: str, request_id: str) -> str:
     with open(filename, "w") as file:  # pylint: disable=W1514
         file.write(data_str)
 
-    return "success"
+    return "your local fides_uploads folder"