+++ tests/ops/conftest.py
@@ -255,6 +255,15 @@ def subject_identity_verification_not_required():
     config.execution.subject_identity_verification_required = original_value
 
 
+@pytest.fixture(autouse=True, scope="function")
+def privacy_request_complete_email_notification_disabled():
+    """Disable request completion email for most tests unless overridden"""
+    original_value = config.notifications.send_request_completion_notification
+    config.notifications.send_request_completion_notification = False
+    yield
+    config.notifications.send_request_completion_notification = original_value
+
+
 @pytest.fixture(scope="session", autouse=True)
 def event_loop():
     try: